[executor-types] move private types back to executor crate
